+++ src/arrowspace_tuner/core/__init__.py
@@ -8,10 +8,8 @@ Import from arrowspace_tuner directly:
     from arrowspace_tuner import StudyConfig, BuildParams  # for power users
 """
 from .config import BuildParams, StudyConfig
-from .graph import fiedler_normalized, gl_to_scipy
+from .graph import fiedler_normalized
 from .objective import build_and_score, make_objective
-
-__version__ = "0.1.0"
 
 __all__ = [
     # config
@@ -19,7 +17,6 @@ __all__ = [
     "StudyConfig",
     # graph
     "fiedler_normalized",
-    "gl_to_scipy",
     # objective
     "build_and_score",
     "make_objective",